Move and rename RoutingTableHelper.
