+++ go/vumitools/account/__init__.py
@@ -1,10 +1,9 @@
 from go.vumitools.account.models import (
-    UserTagPermission, UserAppPermission, UserAccount, RoutingTableHelper,
-    AccountStore, PerAccountStore, GoConnector)
+    UserTagPermission, UserAppPermission, UserAccount,
+    AccountStore, PerAccountStore)
 
 
 __all__ = [
     'UserTagPermission', 'UserAppPermission', 'UserAccount',
-    'RoutingTableHelper', 'AccountStore', 'PerAccountStore',
-    'GoConnector',
+    'AccountStore', 'PerAccountStore',
     ]
